create-db/requirements.txt
pandas
psycopg2-binary
sqlalchemy
openpyxl
